requirements.txt
flask
chromadb
python-dotenv
google-generativeai
numpy
google-genai
